Closes #1485: browser-domains: Move new classes to "autocomplete" sub package.
